Extract server types into sibling *.types.ts files
handlers.ts declared sixteen types before its first line of behaviour, which
made the implementation hard to find. Moves the type declarations out of
every server file that declares more than two of them into a sibling
*.types.ts:

  core/handlers.ts -> core/handlers.types.ts  (16 types)
  core/crypto.ts   -> core/crypto.types.ts    (4 types)
  core/router.ts   -> core/router.types.ts    (3 types)

Files at or under the threshold keep their types inline (storage.ts,
storage-fs.ts and index.ts declare one apiece; the test files two), since
splitting those would add indirection without removing clutter.

Each implementation file re-exports the public types it used to declare, so
existing importers of ./handlers, ./crypto and ./router are unchanged and
this stays a pure move.

Records the convention in CLAUDE.md, including the more-than-two threshold,
so future agents apply it without rediscovering the reasoning.
